+++ src/app/core/state/notifications/index.ts
@@ -1,5 +1,1 @@
-export * from './notifications.actions';
-export * from './notifications.effects';
-export * from './notifications.reducer';
-export * from './notifications.selectors';
 export * from './notifications.state';