+++ src/app/features/settings/data-access/state/index.ts
@@ -1,6 +1,3 @@
 // Settings Store Barrel File
-export * from './settings.actions';
-export * from './settings.effects';
-export * from './settings.reducer';
-export * from './settings.selectors';
 export * from './settings.state';
+export * from './settings.store';